add moaifiddle
I saw the other game related stuff and the lua.vm.js mention and thought this might be handy. It is geared towards the moai game engine but since it is just lua it also works fine for normal lua code. Includes an entire port of the moai game engine to js using emscripten and including lua, an in browser editor, a repl, and a virtual filesystem. Does this sort of fit with what you are after?